Move passwords utils to common/auth/passwords
