+++ packages/core/src/common/auth/index.ts
@@ -1,0 +1,1 @@
+export * from './passwords';
