+++ packages/core/src/common/auth/passwords/index.ts
@@ -1,0 +1,2 @@
+export { hashPassword } from './hash-password';
+export { verifyPassword } from './verify-password';
