+++ packages/core/src/common/index.ts
@@ -1,3 +1,4 @@
+export * from './auth';
 export * from './encoding';
 export * from './hooks';
 export * from './tokens';
