+++ packages/core/src/common/utils/index.ts
@@ -3,9 +3,7 @@ export { displayServerURL } from './display-server-url.util';
 export { escapeProp } from './escape-prop';
 export { escape } from './escape';
 export { getAjvInstance } from './get-ajv-instance';
-export { hashPassword } from './hash-password.util';
 export { isInFile } from './is-in-file.util';
 export { renderError } from './render-error.util';
 export { render, renderToString } from './render.util';
 export { streamToBuffer } from './stream-to-buffer';
-export { verifyPassword } from './verify-password.util';
